peopleSim: revert localTransport heap to Float32 — fixes farmland blowup
The earlier Float32->Float64 change to localTransport's distance heap was
a low-severity "precision fix", but it turned out to be load-bearing. The
Float32 rounding quirk had been pruning the bounded-Dijkstra reach early,
which kept each settlement's farmland catchment small; the whole food /
carrying-capacity balance was tuned around that. Switching to Float64
made the Dijkstra explore the full cost budget, so the catchment floods
along cheap terrain/roads and farmland grew toward its pop/16 target —
roughly tripling farm tiles (probe: 1413 -> 4930) and giving big cities
thousands of tiles.

Reverted to Float32 and documented why it must stay that way unless
MAX_TRANSPORT_BY_TIER is re-tuned. Probe farmland is back to ~1413.

+++ src/peopleSim/settlement.js
@@ -908,19 +908,23 @@ function localTransport(world, sx, sy, maxCost, kn) {
 class _MinHeap {
   constructor(cap = 256) {
     this.ti = new Int32Array(cap);
-    // Float64 distances: localTransport's `out` is a Map (Float64), and
-    // the staleness check `d > out.get(ti)` mis-fires on Float32 rounding
-    // — a value pushed at d=4.24445128 reads back as 4.24445152 and
-    // silently drops the expansion, truncating the reach set. (Same
-    // reason roads.js's MinHeap uses Float64.)
-    this.d  = new Float64Array(cap);
+    // NOTE: Float32 here is DELIBERATE, not an oversight. localTransport's
+    // `out` is a Float64 Map, so the staleness check `d > out.get(ti)` can
+    // mis-fire on Float32 rounding and prune branches early — which keeps
+    // each settlement's farmland catchment small. The food / carrying-
+    // capacity balance is tuned around that smaller catchment; switching
+    // to Float64 (a "correct" Dijkstra) lets the catchment flood along
+    // cheap terrain and roughly triples farmland, and would need
+    // MAX_TRANSPORT_BY_TIER re-tuned to compensate. Don't change without
+    // re-tuning the farmland reach.
+    this.d  = new Float32Array(cap);
     this.n  = 0;
     this.cap = cap;
   }
   _grow() {
     const ncap = this.cap * 2;
     const nti = new Int32Array(ncap); nti.set(this.ti);
-    const nd  = new Float64Array(ncap); nd.set(this.d);
+    const nd  = new Float32Array(ncap); nd.set(this.d);
     this.ti = nti; this.d = nd; this.cap = ncap;
   }
   push(ti, d) {
